backend/requirements.txt
Flask
python-dotenv
pymongo
Flask-Cors
gunicorn
Werkzeug
PyJWT
pytz
openai
